Added project key to the djs.json file.

+++ util/util.js
@@ -7,9 +7,9 @@ const CURRENT_DIR = process.cwd();
  * Generates a DiscordJS project. If framework is enabled, it will set the framework
  * key in djs.json to "commando". 
  */
-module.exports.generateProject = async function(type, projectName, usingFramework, configObj) {
+module.exports.generateProject = function(type, projectName, usingFramework, configObj) {
     if(type === 'djs') {
-        let opts = usingFramework ? { framework: 'commando', groups: [] } : { framework: 'none' }
+        let opts = usingFramework ? { project: projectName, framework: true, groups: [] } : { project: projectName, framework: false }
         fs.mkdir(path.join(CURRENT_DIR, projectName))
         .then(() => fs.mkdir(path.join(CURRENT_DIR, projectName, 'config'))) // Create Config Folder.
         .then(() => usingFramework ? fs.copyFile(path.join(__dirname, '..', 'templates', 'registry.js'), path.join(CURRENT_DIR, projectName, 'config', 'registry.js')) : Promise.resolve())
@@ -23,9 +23,8 @@ module.exports.generateProject = async function(type, projectName, usingFramewor
     }
 }
 
-module.exports.exists = async function(projectName) {
-    let file = path.join(CURRENT_DIR, projectName);
-    console.log(file);
+module.exports.exists = async function(filename) {
+    let file = path.join(CURRENT_DIR, filename);
     try {
         const res = await fs.access(file);
         return Promise.resolve(true); // Resolve true if project directory exists.
@@ -34,6 +33,10 @@ module.exports.exists = async function(projectName) {
         // If fs.access throws an err, that means file doesn't exist.
         return Promise.resolve(false); // Resolve false if project directory doesn't exist.
     }
+}
+
+module.exports.readFile = async function(file) {
+    return await fs.readFile(path.join(CURRENT_DIR, file), 'utf8');
 }
 
 module.exports.generateCommandTemplate = async function(projectName, options) {
@@ -53,10 +56,14 @@ module.exports = class ${options.name}Command extends commando.Command {
     }
 }`
     // First check if the group directory exists.
-    
     let doesExist = await this.exists(path.join('commands', `${options.group}`));
     if(doesExist) {
-        await fs.writeFile(path.join(CURRENT_DIR, 'commands', `${options.group}`, `${options.name}Command.js`), template);
+        let cmdExist = await this.exists(path.join('commands', `${options.group}`, `${options.name}Command.js`));
+        if(!cmdExist) {
+            await fs.writeFile(path.join(CURRENT_DIR, 'commands', `${options.group}`, `${options.name}Command.js`), template);
+        } else {
+            throw new Error("Command already exists under that group.");
+        }
     }
     else {
         await fs.mkdir(path.join(CURRENT_DIR, 'commands', `${options.group}`));
